Refactor out common filter options
