+++ tests/data/productCategories/Shared.ts
@@ -1,0 +1,20 @@
+export type ProductCategoryExpectedText = {
+  Breadcrumbs: string;
+  Title: string;
+  Filters: string[];
+};
+
+export const FilterOptions = [
+  'STYLE',
+  'SIZE',
+  'PRICE',
+  'COLOR',
+  'MATERIAL',
+  'ECO COLLECTION',
+  'PERFORMANCE FABRIC',
+  'ERIN RECOMMENDS',
+  'NEW',
+  'SALE',
+  'PATTERN',
+  'CLIMATE',
+];
