+++ tests/data/productCategories/WomenHoodies.ts
@@ -1,20 +1,7 @@
-import { ProductCategoryExpectedText } from '../productCategoryPage';
+import { FilterOptions, ProductCategoryExpectedText } from './Shared';
 
 export const ExpectedText: ProductCategoryExpectedText = {
   Breadcrumbs: 'Home  Women  Tops  Hoodies & Sweatshirts',
   Title: 'Hoodies & Sweatshirts',
-  Filters: [
-    'STYLE',
-    'SIZE',
-    'PRICE',
-    'COLOR',
-    'MATERIAL',
-    'ECO COLLECTION',
-    'PERFORMANCE FABRIC',
-    'ERIN RECOMMENDS',
-    'NEW',
-    'SALE',
-    'PATTERN',
-    'CLIMATE',
-  ],
+  Filters: [...FilterOptions],
 };
