+++ tests/data/productCategories/WomenShorts.ts
@@ -1,20 +1,7 @@
-import { ProductCategoryExpectedText } from '../productCategoryPage';
+import { FilterOptions, ProductCategoryExpectedText } from './Shared';
 
 export const ExpectedText: ProductCategoryExpectedText = {
   Breadcrumbs: 'Home  Women  Bottoms  Shorts',
   Title: 'Shorts',
-  Filters: [
-    'STYLE',
-    'SIZE',
-    'PRICE',
-    'COLOR',
-    'MATERIAL',
-    'ECO COLLECTION',
-    'PERFORMANCE FABRIC',
-    'ERIN RECOMMENDS',
-    'NEW',
-    'SALE',
-    'PATTERN',
-    'CLIMATE',
-  ],
+  Filters: [...FilterOptions],
 };
