+++ tests/data/productCategories/WomenTops.ts
@@ -1,21 +1,7 @@
-import { ProductCategoryExpectedText } from '../productCategoryPage';
+import { FilterOptions, ProductCategoryExpectedText } from './Shared';
 
 export const ExpectedText: ProductCategoryExpectedText = {
   Breadcrumbs: 'Home  Women  Tops',
   Title: 'Tops',
-  Filters: [
-    'CATEGORY',
-    'STYLE',
-    'SIZE',
-    'PRICE',
-    'COLOR',
-    'MATERIAL',
-    'ECO COLLECTION',
-    'PERFORMANCE FABRIC',
-    'ERIN RECOMMENDS',
-    'NEW',
-    'SALE',
-    'PATTERN',
-    'CLIMATE',
-  ],
+  Filters: ['CATEGORY', ...FilterOptions],
 };
